first version

+++ src/Characters/Character.java
@@ -21,9 +21,9 @@ public class Character {
     private int maxStrength;
     private int maxVitality;
     private int maxIntelligence;
-    private int maxHP;
+    private double maxHP;
 
-    public Character(int maxStrength, int maxVitality, int maxIntelligence, int maxHP) {
+    public Character(int maxStrength, int maxVitality, int maxIntelligence, double maxHP) {
         this.maxStrength = maxStrength;
         this.maxVitality = maxVitality;
         this.maxIntelligence = maxIntelligence;
@@ -87,11 +87,11 @@ public class Character {
         this.maxIntelligence = maxIntelligence;
     }
 
-    public int getMaxHP() {
+    public double getMaxHP() {
         return maxHP;
     }
 
-    public void setMaxHP(int maxHP) {
+    public void setMaxHP(double maxHP) {
         this.maxHP = maxHP;
     }
 
@@ -142,4 +142,5 @@ public class Character {
     public void setArmor(Items armor) {
         this.armor = armor;
     }
+
 }
