+++ src/Items/Shield.java
@@ -1,7 +1,7 @@
 package Items;
 
 public class Shield extends Weapon{
-    public Shield(String name, int weight, int value, int damage) {
+    public Shield(String name, int weight, int value, double damage) {
         super(name, weight, value,damage);
     }
     public void stun(){
